backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
transformers==4.33.2
requests==2.31.0
torch
waitress==2.1.2
